requirements.txt
matplotlib
msvc-runtime
scikit-image==0.18.1
rawpy==0.16.0
numpy
pyyaml
exifread
pywin32
prometheus-client
